Issue number:  CACHE-128, CACHE-38, CACHE-135, CACHE-129, CACHE-83 and CACHE-69 Submitted by:  Lars Torunski [ltorunski@t-online.de] Reviewed by:   Andres March
Unfortunately, it was too difficult to break this into multiple commits because of the way the patch was submitted.

Added a "Expires" value to the CacheHttpServletResponseWrapper to implement CACHE-83. The "Expires" value is a fixed date and time, when the cache entry is going to be stale (default value is infinity).

Now the user can define, if the CacheFilter handles fragments or not. And the user can define that the CacheFilter should detect whether the reponse is included in a page or not. The default parameter is auto detect.
Furthermore I had to correct parameter in the CacheHttpServletResponseWrapper.