Issue number:  CACHE-170 Submitted by:  Guillaume Berche Reviewed by:   Andres March
Had to change the management of the update states to ensure 2 states are not created for the same key, thereby keeping threads from being notified properly. One thing to revisit is the placement of the startUpdate call.  Realistically this should be at the beginning of the put() but I think this will perform better and should not cause problems.